Create Asynchronous Counter.py
This is an asynchronous counter, when building python projects you may find that there are times when you may want to run multiple pieces of code at the same time. For instance say you want load information from a url and your GUI needs to wait a while for the connection session to initiate and for the actual transfer to happen. Due to the fact that python can only run one thread at a time your GUI may freeze while the thread is establishing a session and transfering data. You can however avoid this freezing issue by using the power of the threading module in python. This code goes over how you can create a counter in the background while performing other tasks so they can be timed, Functions can easily be replace to fit whatever purpose you desire.